refactor: separated percentage-selector and percentage-message to its own component
