+++ packages/wallets/src/features/cashier/modules/WithdrawalCrypto/components/WithdrawalCryptoForm/WithdrawalCryptoForm.scss
@@ -5,24 +5,6 @@
     align-items: center;
     gap: 2.4rem;
 
-    &__percentage {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        gap: 0.4rem;
-
-        &-message {
-            width: 100%;
-            height: 1.8rem;
-            text-align: center;
-
-            @include mobile {
-                height: 1.4rem;
-            }
-        }
-    }
-
     &__submit {
         width: 100%;
         display: flex;
