+++ packages/wallets/src/features/cashier/modules/WithdrawalCrypto/components/WithdrawalCryptoForm/WithdrawalCryptoForm.tsx
@@ -1,25 +1,15 @@
 import React from 'react';
 import { Field, FieldProps, Formik } from 'formik';
-import { WalletButton, WalletsPercentageSelector, WalletText, WalletTextField } from '../../../../../../components';
+import { WalletButton, WalletTextField } from '../../../../../../components';
 import { useWithdrawalCryptoValidator } from '../../hooks';
 import { useWithdrawalCryptoContext } from '../../provider/WithdrawalCryptoProvider';
 import { WithdrawalCryptoAmountConverter } from './components/WithdrawalCryptoAmountConverter';
+import { WithdrawalCryptoPercentageSelector } from './components/WithdrawalCryptoPercentageSelector';
 import './WithdrawalCryptoForm.scss';
 
 const WithdrawalCryptoForm: React.FC = () => {
-    const { activeWallet, exchangeRates, fractionalDigits, getConvertedFiatAmount, requestCryptoWithdrawal } =
-        useWithdrawalCryptoContext();
+    const { activeWallet, fractionalDigits, requestCryptoWithdrawal } = useWithdrawalCryptoContext();
     const { validateCryptoAddress } = useWithdrawalCryptoValidator(activeWallet, fractionalDigits);
-
-    const getPercentageMessage = (value: string) => {
-        const amount = parseFloat(value);
-        if (!activeWallet?.balance) return;
-
-        if (amount <= activeWallet?.balance) {
-            const percentage = Math.round((amount * 100) / activeWallet?.balance);
-            return `${percentage}% of available balance ${activeWallet?.display_balance}`;
-        }
-    };
 
     return (
         <Formik
@@ -35,7 +25,7 @@ const WithdrawalCryptoForm: React.FC = () => {
                 })
             }
         >
-            {({ errors, handleSubmit, isSubmitting, setValues, values }) => {
+            {({ errors, handleSubmit, isSubmitting, values }) => {
                 return (
                     <form autoComplete='off' className='wallets-withdrawal-crypto-form' onSubmit={handleSubmit}>
                         <div className='wallets-withdrawal-crypto-address'>
@@ -51,39 +41,7 @@ const WithdrawalCryptoForm: React.FC = () => {
                                 )}
                             </Field>
                         </div>
-                        <div className='wallets-withdrawal-crypto-form__percentage'>
-                            <WalletsPercentageSelector
-                                amount={
-                                    activeWallet?.balance &&
-                                    exchangeRates?.data?.rates &&
-                                    !Number.isNaN(parseFloat(values.cryptoAmount)) &&
-                                    parseFloat(values.cryptoAmount) <= activeWallet.balance
-                                        ? parseFloat(values.cryptoAmount)
-                                        : 0
-                                }
-                                balance={activeWallet?.balance ?? 0}
-                                onChangePercentage={percentage => {
-                                    if (activeWallet?.balance) {
-                                        const fraction = percentage / 100;
-                                        const cryptoAmount = (activeWallet?.balance * fraction).toFixed(
-                                            fractionalDigits.crypto
-                                        );
-                                        const fiatAmount = getConvertedFiatAmount(cryptoAmount);
-
-                                        return setValues({
-                                            ...values,
-                                            cryptoAmount,
-                                            fiatAmount,
-                                        });
-                                    }
-                                }}
-                            />
-                            <div className='wallets-withdrawal-crypto-form__percentage-message'>
-                                <WalletText color='less-prominent' size='xs'>
-                                    {getPercentageMessage(values.cryptoAmount)}
-                                </WalletText>
-                            </div>
-                        </div>
+                        <WithdrawalCryptoPercentageSelector />
                         <WithdrawalCryptoAmountConverter />
                         <div className='wallets-withdrawal-crypto-form__submit'>
                             <WalletButton
