+++ packages/wallets/src/features/cashier/modules/WithdrawalCrypto/components/WithdrawalCryptoForm/components/WithdrawalCryptoPercentageSelector/index.ts
@@ -1,0 +1,1 @@
+export { default as WithdrawalCryptoPercentageSelector } from './WithdrawalCryptoPercentageSelector';
